add: enhance dashboard (home) for mortality reports with new charts and data visualization
